added validation to check if fiance istalled

+++ grails-app/services/net/hedtech/banner/finance/requisition/system/DocumentManagementCompositeService.groovy
@@ -21,10 +21,14 @@ class DocumentManagementCompositeService {
     def bdmAttachmentService
     def requisitionHeaderService
 
-    def uploadDocument( file, requisitionCode, docType, ownerPidm, vpdiCode ) {
+
+    def uploadDocument( file, requisitionCode, docType, ownerPidm, vpdiCode, bdmInstalled ) {
+        if (!bdmInstalled) {
+            throw new ApplicationException( DocumentManagementCompositeService, new BusinessLogicValidationException( FinanceProcurementConstants.ERROR_MESSAGE_BDM_NOT_INSTALLED, [] ) )
+        }
         try {
             def map = bdmAttachmentService.createBDMLocation( file )
-            def requisition = requisitionHeaderService.findRequisitionHeaderByRequestCode(requisitionCode)
+            def requisition = requisitionHeaderService.findRequisitionHeaderByRequestCode( requisitionCode )
             uploadDocToBdmServer( requisition, docType, ownerPidm, map.fileName, map.absoluteFileName, vpdiCode )
             map.userDir.deleteDir()
         } catch (FileNotFoundException e) {
@@ -34,65 +38,73 @@ class DocumentManagementCompositeService {
         }
     }
 
-    def uploadDocToBdmServer(RequisitionHeader requisition, docType, ownerPidm, fileName, absoluteFileName, vpdiCode) throws ApplicationException{
 
-        DateFormat dateFormat = new SimpleDateFormat(FinanceProcurementConstants.BDM_DATE_FORMAT)
+    private
+    def uploadDocToBdmServer( RequisitionHeader requisition, docType, ownerPidm, fileName, absoluteFileName, vpdiCode ) throws ApplicationException {
+
+        DateFormat dateFormat = new SimpleDateFormat( FinanceProcurementConstants.BDM_DATE_FORMAT )
 
         def documentAttributes = [:]
         //This empty string is just to verify all fields data is set
         String emptyString = ''
-        documentAttributes.put(FinanceProcurementConstants.BDM_DOCUMENT_ID,requisition.requestCode)
-        documentAttributes.put(FinanceProcurementConstants.BDM_BANNER_DOC_TYPE,docType)
-        documentAttributes.put(FinanceProcurementConstants.BDM_DOCUMENT_TYPE,docType)
-        documentAttributes.put(FinanceProcurementConstants.BDM_TRANSACTION_DATE,requisition.transactionDate)
-        documentAttributes.put(FinanceProcurementConstants.BDM_VENDOR_ID,requisition?.vendorPidm ?requisition.vendorPidm :emptyString)
-        documentAttributes.put(FinanceProcurementConstants.BDM_VENDOR_NAME,requisition?.vendorPidm ? PersonIdentificationName.findByPidm(requisition?.vendorPidm)?.fullName:emptyString)
-        documentAttributes.put(FinanceProcurementConstants.BDM_FIRST_NAME,fileName)
-        documentAttributes.put(FinanceProcurementConstants.BDM_PIDM,ownerPidm)
-        documentAttributes.put(FinanceProcurementConstants.BDM_ROUTING_STATUS,emptyString)
-        documentAttributes.put(FinanceProcurementConstants.BDM_ACTIVITY_DATE,dateFormat.format(new Date()))
-        documentAttributes.put(FinanceProcurementConstants.BDM_DISPOSITION_DATE,emptyString)
+        documentAttributes.put( FinanceProcurementConstants.BDM_DOCUMENT_ID, requisition.requestCode )
+        documentAttributes.put( FinanceProcurementConstants.BDM_BANNER_DOC_TYPE, docType )
+        documentAttributes.put( FinanceProcurementConstants.BDM_DOCUMENT_TYPE, docType )
+        documentAttributes.put( FinanceProcurementConstants.BDM_TRANSACTION_DATE, requisition.transactionDate )
+        documentAttributes.put( FinanceProcurementConstants.BDM_VENDOR_ID, requisition?.vendorPidm ? requisition.vendorPidm : emptyString )
+        documentAttributes.put( FinanceProcurementConstants.BDM_VENDOR_NAME, requisition?.vendorPidm ? PersonIdentificationName.findByPidm( requisition?.vendorPidm )?.fullName : emptyString )
+        documentAttributes.put( FinanceProcurementConstants.BDM_FIRST_NAME, fileName )
+        documentAttributes.put( FinanceProcurementConstants.BDM_PIDM, ownerPidm )
+        documentAttributes.put( FinanceProcurementConstants.BDM_ROUTING_STATUS, emptyString )
+        documentAttributes.put( FinanceProcurementConstants.BDM_ACTIVITY_DATE, dateFormat.format( new Date() ) )
+        documentAttributes.put( FinanceProcurementConstants.BDM_DISPOSITION_DATE, emptyString )
 
-        bdmAttachmentService.createDocument(getBdmParams(), absoluteFileName, documentAttributes, vpdiCode)
+        bdmAttachmentService.createDocument( getBdmParams(), absoluteFileName, documentAttributes, vpdiCode )
     }
 
 
-    def listDocumentsByRequisitionCode(def requisitionCode, vpdiCode){
-           def criteria = [:]
-           criteria.put(FinanceProcurementConstants.BDM_DOCUMENT_ID, requisitionCode)
-           def documentList
-           try {
-               documentList = bdmAttachmentService.viewDocument(getBdmParams(), criteria, vpdiCode)
-           }catch (ApplicationException ae) {
-               throw ae
-           }
-        def dataMap =[:]
-           def requisition = requisitionHeaderService.findRequisitionHeaderByRequestCode(requisitionCode)
-           def docWithUsername = []
-           SimpleDateFormat dateFormat = new SimpleDateFormat(FinanceProcurementConstants.BDM_DATE_FORMAT)
-           documentList.each{doc->
-               String ownerPidm = doc.docAttributes.get('OWNER_PIDM')
-               String ownerName = PersonIdentificationName.findByPidm(ownerPidm)?.fullName
-               Map docAttrs = doc.docAttributes
-               docAttrs.put('USER_NAME', ownerName)
-               docAttrs.put(FinanceProcurementConstants.BDM_ACTIVITY_DATE, dateFormat.parse(docAttrs[FinanceProcurementConstants.BDM_ACTIVITY_DATE]))
-               doc.docAttributes = docAttrs
-               docWithUsername.add(doc)
-           }
+    def listDocumentsByRequisitionCode( def requisitionCode, vpdiCode, bdmInstalled ) {
+        def criteria = [:]
+        criteria.put( FinanceProcurementConstants.BDM_DOCUMENT_ID, requisitionCode )
+        def documentList
+        try {
+            if (!bdmInstalled) {
+                throw new ApplicationException( DocumentManagementCompositeService, new BusinessLogicValidationException( FinanceProcurementConstants.ERROR_MESSAGE_BDM_NOT_INSTALLED, [] ) )
+            }
+            documentList = bdmAttachmentService.viewDocument( getBdmParams(), criteria, vpdiCode )
+        } catch (ApplicationException ae) {
+            throw ae
+        }
+        def dataMap = [:]
+        def requisition = requisitionHeaderService.findRequisitionHeaderByRequestCode( requisitionCode )
+        def docWithUsername = []
+        SimpleDateFormat dateFormat = new SimpleDateFormat( FinanceProcurementConstants.BDM_DATE_FORMAT )
+        documentList.each {doc ->
+            String ownerPidm = doc.docAttributes.get( 'OWNER_PIDM' )
+            String ownerName = PersonIdentificationName.findByPidm( ownerPidm )?.fullName
+            Map docAttrs = doc.docAttributes
+            docAttrs.put( 'USER_NAME', ownerName )
+            docAttrs.put( FinanceProcurementConstants.BDM_ACTIVITY_DATE, dateFormat.parse( docAttrs[FinanceProcurementConstants.BDM_ACTIVITY_DATE] ) )
+            doc.docAttributes = docAttrs
+            docWithUsername.add( doc )
+        }
         dataMap.documentList = documentList
         return dataMap
-       }
+    }
 
     /**
      * this method will delete the documents uploaded to BDM server
      * @param documentId
      * @param vpdiCode
      */
-    def deleteDocumentsByRequisitionCode(documentId, vpdiCode){
+    def deleteDocumentsByRequisitionCode( documentId, vpdiCode, bdmInstalled ) {
         def docIds = []
-        docIds.add(documentId)
+        docIds.add( documentId )
         try {
-            bdmAttachmentService.deleteDocument(getBdmParams(), docIds, vpdiCode)
+            if (!bdmInstalled) {
+                throw new ApplicationException( DocumentManagementCompositeService, new BusinessLogicValidationException( FinanceProcurementConstants.ERROR_MESSAGE_BDM_NOT_INSTALLED, [] ) )
+            }
+            bdmAttachmentService.deleteDocument( getBdmParams(), docIds, vpdiCode )
         } catch (ApplicationException ae) {
             throw ae
         }
@@ -102,13 +114,13 @@ class DocumentManagementCompositeService {
      *
      * @return
      */
-    private Map getBdmParams(){
+    private Map getBdmParams() {
         Map bdmParams = new HashMap()
-        ConfigurationHolder.config.bdmserver.each{k,v->
-            bdmParams.put(k,v)
+        ConfigurationHolder.config.bdmserver.each {k, v ->
+            bdmParams.put( k, v )
         }
         //bdmParams.put('DataSource',ConfigurationHolder.config.bdmserver.BdmDataSource)
-        log.info("BDMParams :: "+bdmParams)
+        log.info( "BDMParams :: " + bdmParams )
 
         return bdmParams
     }
